+++ apps/frontend-manage/src/components/sessions/creation/MultistepWizard.tsx
@@ -1,11 +1,19 @@
+import {
+  faArrowRight,
+  faCancel,
+  faSave,
+} from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { LearningElementOrderType } from '@klicker-uzh/graphql/dist/ops'
-import { Button, Workflow } from '@uzh-bf/design-system'
+import { Button, H2, Workflow } from '@uzh-bf/design-system'
 import { Form, Formik } from 'formik'
 import { useTranslations } from 'next-intl'
 import React, { useEffect, useState } from 'react'
 import CompletionStep from './CompletionStep'
 
 interface MultistepWizardProps {
+  customCompletionAction?: React.ReactNode
+  title: string
   isCompleted: boolean
   editMode: boolean
   completionSuccessMessage?: (elementName: string) => React.ReactNode
@@ -14,6 +22,7 @@ interface MultistepWizardProps {
   onSubmit: (values: any, bag: any) => void
   onViewElement: () => void
   onRestartForm: () => void
+  onCloseWizard: () => void
   workflowItems: {
     title: string
     tooltip?: string
@@ -77,6 +86,8 @@ function Validator({
 }
 
 function MultistepWizard({
+  customCompletionAction,
+  title,
   children,
   initialValues,
   onSubmit,
@@ -85,6 +96,7 @@ function MultistepWizard({
   completionSuccessMessage,
   onViewElement,
   onRestartForm,
+  onCloseWizard,
   workflowItems,
 }: MultistepWizardProps) {
   const t = useTranslations()
@@ -112,17 +124,22 @@ function MultistepWizard({
   }
 
   return (
-    <>
-      <Formik
-        initialValues={initialValues}
-        onSubmit={handleSubmit}
-        validationSchema={step.props.validationSchema}
-        isInitialValid={editMode}
-        enableReinitialize
-      >
-        {({ values, isSubmitting, isValid, resetForm, validateForm }) => (
-          <Form className="border rounded-md h-76 border-uzh-grey-60">
-            <Validator stepNumber={stepNumber} validateForm={validateForm} />
+    <Formik
+      initialValues={initialValues}
+      onSubmit={handleSubmit}
+      validationSchema={step.props.validationSchema}
+      isInitialValid={editMode}
+      enableReinitialize
+    >
+      {({ values, isSubmitting, isValid, resetForm, validateForm }) => (
+        <Form className="border-b h-76 border-uzh-grey-60">
+          <Validator stepNumber={stepNumber} validateForm={validateForm} />
+          <div className="flex flex-row gap-8 items-end">
+            <H2 className={{ root: 'flex flex-none m-0 items-end' }}>
+              {editMode
+                ? t('manage.questionForms.editElement', { element: title })
+                : t('manage.questionForms.createElement', { element: title })}
+            </H2>
             <Workflow
               items={workflowItems}
               onClick={(_, ix) => setStepNumber(ix)}
@@ -132,41 +149,65 @@ function MultistepWizard({
               minimal
               showTooltipSymbols
               className={{
-                item: 'first:rounded-tl-md last:rounded-tr-md',
+                item: 'last:rounded-tr-md',
               }}
             />
-            <div className="flex flex-col justify-between gap-1 p-4 md:h-60">
-              {!isCompleted && <>{step}</>}
-              {isCompleted && (
-                <CompletionStep
-                  completionSuccessMessage={completionSuccessMessage}
-                  name={values.name}
-                  editMode={editMode}
-                  onViewElement={onViewElement}
-                  onRestartForm={onRestartForm}
-                  resetForm={resetForm}
-                  setStepNumber={setStepNumber}
-                />
-              )}
-              {!isCompleted && (
+          </div>
+
+          <div className="flex flex-col justify-between gap-1 py-4 md:h-60">
+            {!isCompleted && <>{step}</>}
+
+            {isCompleted && (
+              <CompletionStep
+                completionSuccessMessage={completionSuccessMessage}
+                name={values.name}
+                editMode={editMode}
+                onViewElement={onViewElement}
+                onRestartForm={onRestartForm}
+                resetForm={resetForm}
+                setStepNumber={setStepNumber}
+              >
+                {customCompletionAction}
+              </CompletionStep>
+            )}
+
+            {!isCompleted && (
+              <div className="flex flex-row justify-between pt-2">
+                <Button
+                  className={{ root: 'border-red-400' }}
+                  onClick={() => onCloseWizard()}
+                  data={{ cy: 'cancel-session-creation' }}
+                >
+                  <FontAwesomeIcon icon={faCancel} />
+                  <div>
+                    {editMode
+                      ? t('manage.questionForms.cancelEditing')
+                      : t('manage.questionForms.cancelCreation')}
+                  </div>
+                </Button>
                 <Button
                   disabled={isSubmitting || !isValid}
                   type="submit"
                   data={{ cy: 'next-or-submit' }}
                   className={{ root: 'w-max self-end' }}
                 >
+                  {stepNumber === steps.length - 1 ? (
+                    <FontAwesomeIcon icon={faSave} />
+                  ) : (
+                    <FontAwesomeIcon icon={faArrowRight} />
+                  )}
                   {stepNumber === steps.length - 1
                     ? editMode
                       ? t('shared.generic.save')
                       : t('shared.generic.create')
                     : t('shared.generic.continue')}
                 </Button>
-              )}
-            </div>
-          </Form>
-        )}
-      </Formik>
-    </>
+              </div>
+            )}
+          </div>
+        </Form>
+      )}
+    </Formik>
   )
 }