+++ apps/frontend-pwa/worker/index.ts
@@ -34,7 +34,7 @@ self.addEventListener('notificationclick', function (event) {
   event.notification.close()
   event.waitUntil(
     clients
-      .matchAll({ type: 'wind ow', includeUncontrolled: true })
+      .matchAll({ type: 'window', includeUncontrolled: true })
       .then(function (clientList) {
         if (clientList.length > 0) {
           let client = clientList[0]