+++ cypress/cypress/e2e/questions-workflow.cy.ts
@@ -87,6 +87,10 @@ describe('Create questions', () => {
       cy.get('[data-cy="insert-answer-field"]').click().type('50%')
       cy.get('[data-cy="add-new-answer"]').click({ force: true })
       cy.get('[data-cy="insert-answer-field"]').eq(1).click().type('100%')
+      cy.get('[data-cy="add-new-answer"]').click({ force: true })
+      cy.get('[data-cy="insert-answer-field"]').eq(2).click().type('75%')
+      cy.get('[data-cy="add-new-answer"]').click({ force: true })
+      cy.get('[data-cy="insert-answer-field"]').eq(3).click().type('60%')
       cy.get('[data-cy="save-new-question"]').click({ force: true })
 
       cy.contains('[data-cy="question-block"]', questionTitle)
@@ -99,8 +103,7 @@ describe('Create questions', () => {
         .parentsUntil('[data-cy="question-block"]')
         .find('[data-cy="edit-question"]')
         .click()
-      // cy.get('[data-cy="edit-question"]').first().click(); // TODO Risky at the moment, but no problem once we work with empty database before every test
-      cy.get('[data-cy="sc-answer-options"]').should('have.length', 2)
+      cy.get('[data-cy="kp-answer-options"]').should('have.length', 4)
     }),
     it('creates a Numeric question', () => {
       const randomQuestionNumber = Math.round(Math.random() * 1000)