+++ packages/graphql/src/services/sessions.ts
@@ -66,6 +66,9 @@ export function prepareInitialInstanceResults(
     case QuestionType.FREE_TEXT: {
       return {}
     }
+
+    // default:
+    //   return {}
   }
 }
 
@@ -138,6 +141,7 @@ export async function createSession(
               return {
                 order: ix,
                 type: QuestionInstanceType.SESSION,
+                pointsMultiplier: multiplier * question.pointsMultiplier,
                 questionData: processedQuestionData,
                 results: prepareInitialInstanceResults(processedQuestionData),
                 question: {
@@ -270,6 +274,7 @@ export async function editSession(
               return {
                 order: ix,
                 type: QuestionInstanceType.SESSION,
+                pointsMultiplier: multiplier * question.pointsMultiplier,
                 questionData: processedQuestionData,
                 results: prepareInitialInstanceResults(processedQuestionData),
                 question: {
@@ -395,8 +400,6 @@ export async function startSession(
 
       // generate a random pin code
       const pinCode = 100000 + Math.floor(Math.random() * 900000)
-
-      // TODO: if the session is paused, reinitialize and restart
 
       return ctx.prisma.liveSession.update({
         where: {
@@ -1716,10 +1719,10 @@ export async function cancelSession(
     throw new Error('Session is not running')
   }
 
-  const leaderboardEntries = session.blocks
-    .map((block) => block.leaderboard)
-    .flat()
-  const instances = session.blocks.map((block) => block.instances).flat()
+  const leaderboardEntries = session.blocks.flatMap(
+    (block) => block.leaderboard
+  )
+  const instances = session.blocks.flatMap((block) => block.instances)
 
   const [updatedSession] = await ctx.prisma.$transaction([
     ctx.prisma.liveSession.update({
@@ -1776,11 +1779,9 @@ export async function cancelSession(
     }),
 
     ...instances.map((instance) =>
-      ctx.prisma.questionInstance.updateMany({
+      ctx.prisma.questionInstance.update({
         where: {
-          id: {
-            in: instance.id,
-          },
+          id: instance.id,
         },
         data: {
           participants: 0,
@@ -1791,6 +1792,10 @@ export async function cancelSession(
       })
     ),
   ])
+
+  ctx.redisExec.unlink(`s:${id}:meta`)
+  ctx.redisExec.unlink(`s:${id}:lb`)
+  ctx.redisExec.unlink(`s:${id}:xp`)
 
   return updatedSession as ISession
 }