+++ packages/i18n/messages/de.ts
@@ -379,6 +379,11 @@ Da die KlickerUZH-App noch nicht im iOS-App-Store verfügbar ist, folgen Sie die
         'Sie haben bisher keine Fragen gebookmarked. Klicken Sie hierfür einfach auf das Lesezeichen-Symbol auf einer Frage.',
       awards: 'Auszeichnungen',
       open: 'offen',
+      leaveCourseTitle: 'Leaderboard verlassen',
+      leaveCourseConfirmation:
+        'Möchten Sie das Leaderboard wirklich verlassen?',
+      leaveCourseInformation:
+        'Wenn Sie das Kurs-Leaderboard verlassen, werden Interaktionen mit Kursaktivitäten und Ihre gesammelten Punkte gelöscht. Sie können dem Leaderboard jederzeit wieder beitreten, beginnen dann aber von Neuem.',
     },
     joinCourse: {
       title: 'Kurs "{name}" beitreten',
@@ -883,6 +888,10 @@ Da die KlickerUZH-App noch nicht im iOS-App-Store verfügbar ist, folgen Sie die
       NRPrecision: 'Die Anzahl Nachkommastellen muss mindestens 0 sein.',
       solutionRangeRequired:
         'Bitte geben Sie mindestens einen Lösungsbereich an.',
+      NumberQuestionsRequired:
+        'Es muss mindestens eine Antwortmöglichkeit gegeben werden',
+      NumberQuestionsRequiredKPRIM:
+        'Bei Kprim-Fragen müssen genau vier Antwortmöglichkeiten gegeben werden',
     },
     sessions: {
       runningSessions: 'Laufende Sessionen',